third_party/blink/web_tests/platform/generic/fast/events/keyboardevent-repeat-expected.txt

Tests KeyboardEvent.repeat attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lastKeyboardEvent.repeat is false
PASS lastKeyboardEvent.repeat is true
PASS successfullyParsed is true

TEST COMPLETE
